+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Add validation for `--validation` flag value
+
 ## [0.0.1] - 2025-09-29
 
 ### Added